BE_AI_DPA_APP/requirements.txt
hdbcli
generative-ai-hub-sdk[all]
langchain
python-dotenv
# ipywidgets
sqlalchemy-hana
flask
